adding a useful way to validate the json response against a gold standard of the json that we expect.
json and dictionaries do not preserve the order of objects, nor do they have them sorted. From local testing, I was able to see this behavior. Json validation using python cmp() or "==" or even assertDictEqual (which provides useful error messaging on the diff; https://kite.com/python/docs/unittest.TestCase.assertDictEqual) do not have options to sort. We could roll our own dict sorting, using ordered dict (which preserve the order things are added to the dict) or a sorted list. Or we could sort using json.dumps which outputs a long messy string. But then I realized that some of the APIs return codes that are generated every time the API is run. Therefore a direct comparison of the json would not work.

We want to be able to validate some parts of the json exactly (like wanting OK status == True) while having flexibility to specify that we should receive a string for a generated code. That's where Json Schema comes in. https://python-jsonschema.readthedocs.io/en/latest/ It provides the flexibility that we need. Having to use enums is admittedly a bit funky, but it does work for what we want to test, ensuring we get back the values we're expecting.

In order to create new Json Schemas from a json response, use this tool to generate, customize, and validate schemas: https://jsonschema.net/