CA-648 Abstract out Fence JWT Token Storage & Locking (#91)
Pull out the FenceTokenVendingMachine's logic for interacting with Datastore into a separate class, FenceTokenStorage.

This will allow us to inject a stub for unit testing when we cannot use real Datastore or an emulator. We can no longer rely on testbed's datastore stub as it is not available in python 3 app engine.

* Change condition for ServiceAccountNotUpdatedException to raise that function if there is no expired_at message, i.e. it's still locked, instead of only if the updated account does not have a valid expires_at value.
Before, if the update lock expired during the _wait_for_update loop, we could return the FenceServiceAccount without a received key_json, as the update locked FenceServiceAccount does not have any value for expires_at. Now, the expired update locked value raises the exception.