[CA-590] json response verification for integration tests (#92)
* adding a useful way to validate the json response against a gold standard of the json that we expect.

We want to be able to validate some parts of the json exactly (like wanting OK status == True) while having flexibility to specify that we should receive a string for a generated code. That's where Json Schema comes in. https://python-jsonschema.readthedocs.io/en/latest/ It provides the flexibility that we need. Having to use enums is admittedly a bit funky, but it does work for what we want to test, ensuring we get back the values we're expecting.

In order to create new Json Schemas from a json response, use this tool to generate, customize, and validate schemas: https://jsonschema.net/

* Make local envs point at mocks. In order to use integration test servers, we need this completed: https://broadworkbench.atlassian.net/browse/CA-303 .