CA-648 Use Datastore emulator for testing (#97)
Use Datastore emulator for testing.
Set up some integration tests using the Datastore emulator for testing interactions with Datastore. Hook these up to be run by circleci. Add instructions for running datastore emulator tests locally.

With the move to python3, the testbed datastore stub is being removed so we need to use the emulator for testing.